refactor(keybindings): extract the keybinding model into arto-keybindings
Shortcut parsing, the binding sets, the presets, the actions, the context
rules, the matching engine and hint formatting had no need for a window,
yet they lived inside the Dioxus app: `cargo test` on them linked the
whole desktop stack, and `config` and `keybindings` referred to each
other (`config::BindingSet` one way, `keybindings::default_bindings` the
other), so neither could be understood alone.

Move them into a library crate that depends only on keyboard-types,
serde and thiserror. keyboard-types is the crate Dioxus and muda both
re-export their `Key`, `Code` and `Modifiers` from, so chords built in
the crate compare equal to the app's events without any conversion.
`BindingSet` now lives with the engine that consumes it and the app's
`Config` embeds it, which makes the dependency one-directional.

What stays in the app is exactly the glue that needs the app:

- `chord_from_event` builds a chord from a Dioxus `KeyboardEvent` via the
  new `KeyChord::new`, which normalizes the same way the parser does (the
  `From<KeyboardEvent>` impls could not move, both types being foreign);
- `accelerator_for_key` turns the crate's `single_chord_code` into a muda
  `Accelerator`;
- the `shortcut_hint_for_*` helpers read the user's bindings from CONFIG
  and call the crate's `hint_for_action`;
- `engine_for` builds the engine once with `KeybindingEngine::build` and
  logs the entries it had to skip.

That last point is the one behavior change: the library no longer logs.
Entries with an invalid key or an unknown action come back as a
`BindingError` list from `BindingSet::resolve`, and the app decides to
warn about them, once per engine build rather than on every key press.
