+++ crates/arto-keybindings/src/presets/preset.schema.json
