+++ crates/arto/src/components/app/keybinding_engine.rs
@@ -104,13 +104,13 @@ pub(super) fn setup_keybinding_engine(
 ) {
     use crate::config::{CONFIG, CONFIG_CHANGED_BROADCAST};
     use crate::keybindings::dispatcher::dispatch_action;
-    use crate::keybindings::{Action, KeyContext, KeyMatchResult, KeybindingEngine};
-    use crate::shortcut::KeyChord;
+    use crate::keybindings::KeyChord;
+    use crate::keybindings::{Action, KeyContext, KeyMatchResult};
     use std::cell::RefCell;
 
     // use_signal must be called at component render level (not inside use_hook)
     let initial_config = CONFIG.read().keybindings.clone();
-    let engine = use_signal(|| RefCell::new(KeybindingEngine::new(&initial_config)));
+    let engine = use_signal(|| RefCell::new(crate::keybindings::engine_for(&initial_config)));
 
     // spawn/spawn_forever wrapped in use_hook to run only once
     use_hook(move || {
@@ -226,7 +226,7 @@ pub(super) fn setup_keybinding_engine(
             let mut rx = CONFIG_CHANGED_BROADCAST.subscribe();
             while rx.recv().await.is_ok() {
                 let new_config = CONFIG.read().keybindings.clone();
-                *engine.read().borrow_mut() = KeybindingEngine::new(&new_config);
+                *engine.read().borrow_mut() = crate::keybindings::engine_for(&new_config);
                 push_menu_accelerators_to_js();
                 push_reserved_key_overrides_to_js();
                 tracing::debug!("Keybinding engine rebuilt after config change");
