+++ crates/arto/src/config/app_config.rs
@@ -4,7 +4,6 @@ use serde::{Deserialize, Serialize};
 mod behavior;
 mod directory_config;
 mod file_open_behavior;
-mod keybindings_config;
 mod right_sidebar_config;
 mod sidebar_config;
 mod theme_config;
@@ -13,10 +12,10 @@ mod window_position_config;
 mod window_size_config;
 mod zoom_config;
 
+pub use arto_keybindings::{BindingSet, KeyAction};
 pub use behavior::{NewWindowBehavior, StartupBehavior};
 pub use directory_config::DirectoryConfig;
 pub use file_open_behavior::FileOpenBehavior;
-pub use keybindings_config::{BindingSet, KeyAction};
 pub use right_sidebar_config::{RightSidebarConfig, DEFAULT_RIGHT_SIDEBAR_WIDTH};
 pub use sidebar_config::SidebarConfig;
 pub use theme_config::ThemeConfig;
